feat(56): creation d'offre sur un produit (#66)
* feat(56): creation d'offre sur un produit

* feat(56): changement de int pour double

* feat(56): tests

* feat(56): finition des tests

* feat(56): code review

* feat(56): put factory in domain layer

* feat(56): changement de la validation des numeros de telephone

+++ src/main/java/ulaval/glo2003/seller/api/SellerProductResponse.java
@@ -9,7 +9,7 @@ public class SellerProductResponse {
   public String createdAt;
   public String title;
   public String description;
-  public int suggestedPrice;
+  public Double suggestedPrice;
   public OffersResponse offers;
   public List<String> categories;
 
@@ -21,7 +21,7 @@ public class SellerProductResponse {
           String createdAt,
           String title,
           String description,
-          int suggestedPrice,
+          Double suggestedPrice,
           OffersResponse offers,
           List<String> categories
   ) {
